+++ src/main/java/veterinaria/services/impl/UsuarioServiceImpl.java
@@ -6,10 +6,14 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import veterinaria.dto.UsuarioDTO;
 import veterinaria.mappers.UsuarioMapper;
+import veterinaria.models.HistoriaClinica;
 import veterinaria.models.Mascota;
 import veterinaria.models.Usuario;
+import veterinaria.repositorys.HistoriaClinicaRepository;
 import veterinaria.repositorys.MascotaRepository;
 import veterinaria.repositorys.UsuarioRepository;
+import veterinaria.services.HistoriaClinicaService;
+import veterinaria.services.MascotaService;
 import veterinaria.services.UsuarioService;
 
 @Service
@@ -21,7 +25,16 @@ public class UsuarioServiceImpl implements UsuarioService{
     private MascotaRepository mascotaRepository;
     
     @Autowired
+    private HistoriaClinicaRepository historiaRepository;
+    
+    @Autowired
     private UsuarioMapper usuarioMapper;
+    
+    @Autowired
+    private MascotaService mascotaService;
+    
+    @Autowired
+    private HistoriaClinicaService historiaService;
     
     @Override
     public List<UsuarioDTO> obtenerTodo() {
@@ -72,6 +85,22 @@ public class UsuarioServiceImpl implements UsuarioService{
     public void eliminar(Integer id) {
         
         Optional<Usuario> usuarioOptional = usuarioRepository.findById(id);
+        if(!usuarioOptional.isPresent()) return;
+        
+        List<Mascota> mascotas = mascotaRepository.findAll();
+        
+        for(Mascota mascota : mascotas){
+            
+           mascotaService.eliminar(mascota.getId());
+           
+           List<HistoriaClinica> historiasClinicas = historiaRepository.findAll();
+           
+           for(HistoriaClinica historia : historiasClinicas){
+               
+               historiaService.eliminar(historia.getId());
+               
+           }
+        }
         
         usuarioRepository.delete(usuarioOptional.get());
         mascotaRepository.deleteByMascotaId(id);
